requirements.txt
openai-whisper
sounddevice
SpeechRecognition

# Text to Speech
TTS
pydub